read from clm files

+++ src/openmath/glyph.cpp
@@ -16,15 +16,15 @@ const Math Math::empty(0);
 
 int16 KernRecord::operator[](uint16 glyph) const {
   const int i = binSearchIndex(_count, [&](int i) { return glyph - _fields[i << 1]; });
-  return i < 0 ? 0 : _fields[i << 1 + 1];
+  return i < 0 ? 0 : (int16)_fields[i << 1 + 1];
 }
 
 uint16 MathKern::indexOf(int32 height) const {
   if (_count == 0) return 0;
   return binSearchIndex(
-      _count,
-      [&](int i) { return height - _fields[i << 1]; },
-      true);
+    _count,
+    [&](int i) { return height - _fields[i << 1]; },
+    true);
 }
 
 MathKernRecord::MathKernRecord(uint16 ignore) {
